Move AutoEncoder to dxtbx.util
